requirements.txt
python-dotenv
streamlit
langchain
langchain-groq
langchain-core
langchain-community
groq
newsapi-python